03-ballcollisions2d/pyproject.toml
[project]
name = "BallCollisions2D"
version = "0.1.0"
description = "demos from 10 minute physics"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["pyside6>=6.9.0", "ncca-ngl"]
